requirements.txt
# Core dependencies
Flask==3.0.0
requests==2.31.0
redis==5.0.1
python-dotenv==1.0.0

# Security
cryptography==42.0.2
PyJWT==2.8.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.12.0

# Documentation
markdown==3.5.1
mdx_truly_sane_lists==1.3

# Monitoring
prometheus-client==0.19.0
statsd==4.0.1

# Utilities
python-dateutil==2.8.2
pytz==2024.1
aiohttp==3.9.1
backoff==2.2.1

# Development
black==24.1.1
pylint==3.0.3
mypy==1.8.0

flask-socketio==5.3.6
pytest-asyncio==0.23.5
hypothesis==6.98.0
flasgger==0.9.7.1
psutil==5.9.8
tqdm==4.66.1
typing-extensions==4.9.0
python-json-logger==2.0.7
websockets==12.0 